+++ man-roxygen/example.R
@@ -1,20 +1,19 @@
 <%
 pkgs = setdiff(mlr3::lrn(id)$packages, c("mlr3", "mlr3learners"))
-if ("LearnerClassif" %in% class(lrn(id))) {
-  task_id = "sonar"
-} else if ("LearnerRegr" %in% class(lrn(id))) {
-  task_id = "mtcars"
-} else if ("LearnerSurv" %in% class(lrn(id))) {
-  task_id = "grace"
-} else if ("LearnerDens" %in% class(lrn(id))) {
-  task_id = "faithful"
+l  = lrn(id)
+task_id = if ("LearnerClassif" %in% class(l)) {
+  "sonar"
+} else if ("LearnerRegr" %in% class(l)) {
+  "mtcars"
+} else if ("LearnerSurv" %in% class(l)) {
+  "grace"
+} else if ("LearnerDens" %in% class(l)) {
+  "faithful"
 }
 %>
 #'
-#' @examples
-#' set.seed(123)
-#' <%= sprintf("if (mlr3misc::require_namespaces(lrn(\"%s\")$packages, quietly = TRUE))", id) %> {
-#' # Define the Learner and set parameter values
+#' <%= sprintf("@examplesIf mlr3misc::require_namespaces(lrn(\"%s\")$packages, quietly = TRUE)", id) %>
+#' # Define the Learner
 #' <%= sprintf("learner = mlr3::lrn(\"%s\")", id)%>
 #' print(learner)
 #'
@@ -27,15 +26,11 @@ if ("LearnerClassif" %in% class(lrn(id))) {
 #' # Train the learner on the training ids
 #' learner$train(task, row_ids = ids$train)
 #'
-#' # print the model
 #' print(learner$model)
-#'
-#' # importance method
-#' if("importance" %in% learner$properties) print(learner$importance)
+#' <%= if("importance" %in% l$properties) "print(learner$importance)" %>
 #'
 #' # Make predictions for the test rows
 #' predictions = learner$predict(task, row_ids = ids$test)
 #'
 #' # Score the predictions
 #' predictions$score()
-#' }